modified:   General/CellProfiler/CPsubfunctions/CPrelateobjects.m modified:   General/CellProfiler/Modules/RelateAndJoinSegmentation.m
------------------------------------------------------------------------
r1588 | mherrmann | 2014-03-03 11:58:29 +0100 (Mon, 03 Mär 2014) | 2 lines

I found a corpse in the CP basement: stored objects in the handles were identified using 'strfind.m'. This approach is totally unspecific and identifies every object that contains the searchstring, which can be fatal when objects have similar names. I fixed this by replacing 'strfind.m' with 'strcmp.m'. This fix required a hack of 'CPrelateobjects.m', since the names of children were stored with the addition of "counts", which interfered with the use of 'strcmp.m'.